Update : VC

+++ GRAMO/GRAMO/VC/SideMenu/SideMenuViewController.swift
@@ -20,59 +20,59 @@ final class SideMenuViewController: UIViewController {
         majorLabel?.text = UserDefaults.standard.string(forKey: "major")
     }
     
-    @IBAction func calendar(_ sender: UIButton) {
+    @IBAction private func calendar(_ sender: UIButton) {
         let sub = UIStoryboard(name: "Calendar2", bundle: nil)
         let calendar = sub.instantiateViewController(withIdentifier: "Calendar2VC")
         self.navigationController?.pushViewController(calendar, animated: false)
     }
     
-    @IBAction func info(_ sender : UIButton) {
+    @IBAction private func info(_ sender : UIButton) {
         let sub = UIStoryboard(name: "info", bundle: nil)
         let calendar = sub.instantiateViewController(withIdentifier: "infoListViewController")
         self.navigationController?.pushViewController(calendar, animated: false)
     }
     
-    @IBAction func homework(_ sender: UIButton) {
+    @IBAction private func homework(_ sender: UIButton) {
         let sub = UIStoryboard(name: "homework", bundle: nil)
         let calendar = sub.instantiateViewController(withIdentifier: "homeworkVC")
         self.navigationController?.pushViewController(calendar, animated: false)
     }
     
-    @IBAction func logOut(_ sender: UIButton){
-        showDeleteAlert(title: "로그아웃 하시겠습니까?", action: { [self](action) in logout() }, message: nil)
+    @IBAction private func logOut(_ sender: UIButton){
+        showDeleteAlert(title: "로그아웃 하시겠습니까?", action: { [unowned self](action) in logout() }, message: nil)
     }
     
-    @IBAction func withdrawal(_ sender: UIButton){
-        HTTPClient().delete(url: AuthAPI.withDrawel.path(), params: nil, header: Header.token.header()).responseJSON(completionHandler: { res in
+    @IBAction private func withdrawal(_ sender: UIButton){
+        HTTPClient().delete(url: AuthAPI.withDrawel.path(), params: nil, header: Header.token.header()).responseJSON(completionHandler: { [unowned self]res in
             switch res.response?.statusCode {
             case 204 :
                 let sub = UIStoryboard(name: "Auth", bundle: nil)
                 let calendar = sub.instantiateViewController(withIdentifier: "LoginVC")
-                self.navigationController?.pushViewController(calendar, animated: false)
+                navigationController?.pushViewController(calendar, animated: false)
                 
             case 401: print("401 - could not find token user")
-                self.showAlert(title: "허가되지 않은 요청입니다.",message: nil)
+                showAlert(title: "허가되지 않은 요청입니다.",message: nil)
                 
             default:
                 print(res.response?.statusCode ?? "default")
-                self.showAlert(title: "오류가 발생했습니다.", message: nil)
+                showAlert(title: "오류가 발생했습니다.", message: nil)
             }
         })
     }
     
-    func logout() {
-        HTTPClient().delete(url: AuthAPI.logout.path(), params: nil, header: Header.token.header()).responseJSON(completionHandler: { res in
+    private func logout() {
+        HTTPClient().delete(url: AuthAPI.logout.path(), params: nil, header: Header.token.header()).responseJSON(completionHandler: { [unowned self]res in
             switch res.response?.statusCode {
             case 204 :
                 let sub = UIStoryboard(name: "Auth", bundle: nil)
                 let calendar = sub.instantiateViewController(withIdentifier: "LoginVC")
-                self.navigationController?.pushViewController(calendar, animated: false)
+                navigationController?.pushViewController(calendar, animated: false)
                 
             case 401: print("401 - could not find token user")
-                self.showAlert(title: "허가되지 않은 요청입니다.", message: nil)
+                showAlert(title: "허가되지 않은 요청입니다.", message: nil)
             default:
                 print(res.response?.statusCode ?? "default")
-                self.showAlert(title: "오류가 발생했습니다.", message: nil)
+                showAlert(title: "오류가 발생했습니다.", message: nil)
             }
         })
     }
